Extract interface for azure client.
